05_Authentication && 1:N model relation

+++ posts/views.py
@@ -1,13 +1,17 @@
 from django.shortcuts import render, redirect, get_list_or_404, get_object_or_404
-from .forms import PostForm, ImageForm
+from .forms import PostForm, ImageForm, CommentForm
 from .models import Post, Image
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
+@login_required
 def create(request):
     if request.method == 'POST':
         post_form = PostForm(request.POST)
         if post_form.is_valid():
-            post = post_form.save() #게시글 내용 처리 끝
+            post = post_form.save(commit=False) #게시글 내용 처리 끝
+            post.user = request.user
+            post.save()
             for image in request.FILES.getlist('file'):
                 request.FILES['file'] = image
                 image_form = ImageForm(files=request.FILES)
@@ -28,11 +32,18 @@ def create(request):
     
 def list(request):
     posts = get_list_or_404(Post.objects.order_by('-pk'))
-    context = {'posts': posts}
-    return render(request, 'posts/list.html', context)
+    form = CommentForm()
+    context = {'posts': posts,
+                'form': form,
+    }
     
+    return render(request, 'posts/list.html', context)
+
+@login_required    
 def update(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
+    if post.user != request.user:
+        return redirect('posts:list')
     if request.method == 'POST':
         post_form = PostForm(request.POST, instance=post)
         if post_form.is_valid():
@@ -42,9 +53,12 @@ def update(request, post_pk):
         post_form = PostForm(instance=post)
     context = {'post_form': post_form}
     return render(request, 'posts/form.html', context)
-    
+
+@login_required    
 def delete(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
+    if post.user != request.user:
+        return redirect('posts:list')
     if request.method =='POST':
         post.delete()
     return redirect('posts:list')
